issue-86: Fixed copy function, added default image and added a check so it doesn't get deleted, changed seeddata to use default image

+++ Web/CmsControllers/CandidatesController.cs
@@ -513,8 +513,10 @@ namespace Web
                     model.Image.CopyTo(newFile);
                     newFile.Dispose();
                     model.Candidate.Picture = nameOfile;
-                    if (System.IO.File.Exists(oldFileName))
+                    if (System.IO.File.Exists(oldFileName) && oldFileName != "wwwroot\\images/default.jpg")
                     {
+                        Console.WriteLine("TESTING: " + oldFileName);
+                        Console.WriteLine("TESTING: " + "wwwroot\\images/default.jpg");
                         System.IO.File.Delete(oldFileName);
                     }
